requirements.txt
imageio-ffmpeg
matplotlib
numpy 
opencv-python
pandas
Pillow
PyYAML
scikit-image
scikit-learn
scikit-video
scipy
six
sklearn
torch
torchvision
tqdm
wandb
imageio-ffmpeg
av
compressai
psnr-hvm
piq
lpips

